feat: adiciona linkpage do sistema
