+++ src/app/auth/callback/route.ts
@@ -4,7 +4,7 @@ import { getSupabase } from '@/services/supabaseServer';
 export async function GET(request: Request) {
   const { searchParams, origin } = new URL(request.url);
   const code = searchParams.get('code');
-  const next = searchParams.get('next') ?? '/';
+  const next = searchParams.get('next') ?? '/mural';
 
   if (code) {
     const supabase = await getSupabase();
